back full

+++ src/agents/cost_calculator.py
@@ -469,7 +469,7 @@ class GlobalCostCalculator:
 
         # Burnout risk (0-10 scale)
         # High complexity + low team experience = burnout risk
-        burnout_risk = min(10, complexity_multiplier * 2)  # TODO complexity_multiplier * 2 Expected type 'int' (matched generic type 'SupportsRichComparisonT ≤: SupportsDunderLT | SupportsDunderGT'), got 'float' instead
+        burnout_risk = min(10, int(complexity_multiplier * 2))
         if team_velocity.experience == TeamExperience.JUNIOR:
             burnout_risk *= 1.5
         burnout_risk = min(10, burnout_risk)
